done hw7

+++ scrapper/src/main/java/edu/java/scrapper/controllers/LinkController.java
@@ -5,7 +5,12 @@ import edu.java.scrapper.dtos.AddLinkRequestDto;
 import edu.java.scrapper.dtos.LinkResponseDto;
 import edu.java.scrapper.dtos.ListLinkResponseDto;
 import edu.java.scrapper.dtos.RemoveLinkRequestDto;
+import edu.java.scrapper.exceptions.BadLinkException;
+import edu.java.scrapper.exceptions.TooManyRequestsException;
 import edu.java.scrapper.services.LinkService;
+import edu.java.scrapper.utils.BucketGrabber;
+import edu.java.scrapper.utils.LinkResolver;
+import io.github.bucket4j.Bucket;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
@@ -25,30 +30,36 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class LinkController {
     private final LinkService linkService;
+    private final BucketGrabber bucketGrabber;
 
     @GetMapping("/links")
     public ResponseEntity<ListLinkResponseDto> getLinks(
         @RequestHeader("Tg-Chat-Id") Long tgChatId
     ) {
-        List<LinkDto> linkDtoList = linkService.listAll(tgChatId);
-        ListLinkResponseDto listLinkResponseDto = new ListLinkResponseDto();
-        listLinkResponseDto.setLinks(
-            linkDtoList.stream()
-                .map(link -> {
-                    try {
-                        return new LinkResponseDto(link.getId().longValue(), new URI(link.getUrl()));
-                    } catch (URISyntaxException e) {
-                        throw new RuntimeException(e);
-                    }
-                })
-                .toList()
-        );
-        listLinkResponseDto.setSize(listLinkResponseDto.getLinks().size());
-        log.info("Links has been received");
-        return new ResponseEntity<>(
-            listLinkResponseDto,
-            HttpStatus.OK
-        );
+        Bucket bucket = bucketGrabber.grabBucket(tgChatId);
+        if (bucket.tryConsume(1)) {
+            List<LinkDto> linkDtoList = linkService.listAll(tgChatId);
+            ListLinkResponseDto listLinkResponseDto = new ListLinkResponseDto();
+            listLinkResponseDto.setLinks(
+                linkDtoList.stream()
+                    .map(link -> {
+                        try {
+                            return new LinkResponseDto(link.getId().longValue(), new URI(link.getUrl()));
+                        } catch (URISyntaxException e) {
+                            throw new RuntimeException(e);
+                        }
+                    })
+                    .toList()
+            );
+            listLinkResponseDto.setSize(listLinkResponseDto.getLinks().size());
+            log.info("Links has been received");
+            return new ResponseEntity<>(
+                listLinkResponseDto,
+                HttpStatus.OK
+            );
+        } else {
+            throw new TooManyRequestsException();
+        }
     }
 
     @PostMapping("/links")
@@ -56,17 +67,24 @@ public class LinkController {
         @RequestHeader("Tg-Chat-Id") Long tgChatId,
         @RequestBody AddLinkRequestDto addLinkRequestDto
     ) {
-        //TODO date logic
-        linkService.add(tgChatId, addLinkRequestDto.getLink());
-        LinkResponseDto linkResponseDto = new LinkResponseDto(
-            tgChatId,
-            addLinkRequestDto.getLink()
-        );
-        log.info("link has been added");
-        return new ResponseEntity<>(
-            linkResponseDto,
-            HttpStatus.OK
-        );
+        Bucket bucket = bucketGrabber.grabBucket(tgChatId);
+        if (bucket.tryConsume(1)) {
+            //TODO date logic
+            if (!LinkResolver.isSuitableLink(addLinkRequestDto.getLink())) {
+                throw new BadLinkException();
+            }
+            linkService.add(tgChatId, addLinkRequestDto.getLink());
+            LinkResponseDto linkResponseDto = new LinkResponseDto(
+                tgChatId,
+                addLinkRequestDto.getLink()
+            );
+            log.info("link has been added");
+            return new ResponseEntity<>(
+                linkResponseDto,
+                HttpStatus.OK
+            );
+        }
+        throw new TooManyRequestsException();
     }
 
     @DeleteMapping("/links")
@@ -74,15 +92,19 @@ public class LinkController {
         @RequestHeader("Tg-Chat-Id") Long tgChatId,
         @RequestBody RemoveLinkRequestDto removeLinkRequestDto
     ) {
-        linkService.remove(tgChatId, removeLinkRequestDto.getLink());
-        LinkResponseDto linkResponseDto = new LinkResponseDto(
-            tgChatId,
-            removeLinkRequestDto.getLink()
-        );
-        log.info("link has been deleted");
-        return new ResponseEntity<>(
-            linkResponseDto,
-            HttpStatus.OK
-        );
+        Bucket bucket = bucketGrabber.grabBucket(tgChatId);
+        if (bucket.tryConsume(1)) {
+            linkService.remove(tgChatId, removeLinkRequestDto.getLink());
+            LinkResponseDto linkResponseDto = new LinkResponseDto(
+                tgChatId,
+                removeLinkRequestDto.getLink()
+            );
+            log.info("link has been deleted");
+            return new ResponseEntity<>(
+                linkResponseDto,
+                HttpStatus.OK
+            );
+        }
+        throw new TooManyRequestsException();
     }
 }
